requirements.txt
selenium==3.141.0
telegram==0.0.1
python_telegram_bot_calendar==1.0.5
email_validator==1.1.2
mysql_connector_repackaged==0.3.1
mysql-connector==2.2.9
python-telegram-bot==13.4.1
